WIP: Declare a model to record XML-RPC calls
TODO: use the old model and app name, but do so *AFTER*
Kobo has been unloaded from the Django app. Otherwise it
messes up with the migrations. Migration is named
0002_xmlrpcrecord.py and is saved under the kobo directory

this is taken from kobo because we want to get rid of that
dependency and only use a few bits from it.